timesheet-lite/backend/requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
python-jose[cryptography]
python-multipart
pydantic-settings
bcrypt
apscheduler
argon2-cffi
